pyproject.toml
[project]
name = "pypond"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.12"
dependencies = [
    "beartype>=0.19.0",
    "deltalake>=0.25.5",
    "dill>=0.4.0",
    "fastapi>=0.116.1",
    "fsspec>=2024.12.0",
    "kaleido==0.2.1",
    "lancedb>=0.17.0",
    "laspy[lazrs]>=2.5.4",
    "marimo[sql]>=0.14.10",
    "pandas>=2.2.3",
    "parse>=1.20.2",
    "pillow>=11.1.0",
    "plotly>=6.0.0",
    "pydantic-to-pyarrow>=0.1.5",
    "pyiceberg[pyarrow,s3fs,sql-sqlite]>=0.8.1",
    "pytest>=8.3.4",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "sf-hamilton[sdk,ui]>=1.88.0",
    "uvicorn>=0.35.0",
]

[tool.pytest.ini_options]
testpaths = ["tests", "docs"]
addopts = "--codeblocks --ignore=docs/TESTING_GUIDE.md"

[tool.ruff]
src = ["pond", "tests", "conf"]
line-length = 88
target-version = "py310"
extend-exclude = ["docs/*"]

[dependency-groups]
dev = [
    "mkdocs-material>=9.6.16",
    "mkdocstrings[python]>=0.30.0",
    "mypy>=1.16.1",
    "pytest-codeblocks>=0.17.0",
    "types-pyyaml>=6.0.12.20250516",
]
